+++ packages/react-ui/src/theme/sections/components/forms.js
@@ -1,13 +1,18 @@
 import { getSpacing } from '../../themeUtils';
 import { commonPalette } from '../palette';
+import { themeShadows } from '../shadows';
 import { themeTypography } from '../typography';
+
+const switchSizeS = 2;
+const switchSizeM = 3;
+const switchSizeL = 4;
 
 const checkboxRadioOverrides = {
   root: ({ ownerState }) => ({
     padding: getSpacing(0.5),
 
     ...(ownerState.size === 'small' && {
-      padding: getSpacing(0.25)
+      padding: '3px' // Forced to a non-standard value to meet with design
     }),
 
     '&:hover, &:focus-visible': {
@@ -26,7 +31,7 @@ const checkboxRadioOverrides = {
       fontSize: getSpacing(3),
 
       ...(ownerState.size === 'small' && {
-        fontSize: getSpacing(2.5)
+        fontSize: getSpacing(2.25)
       })
     }
   })
@@ -219,133 +224,96 @@ export const formsOverrides = {
 
   // Switch
   MuiSwitch: {
+    defaultProps: {
+      disableRipple: true
+    },
+
     styleOverrides: {
       root: {
-        height: getSpacing(4.5),
-        width: getSpacing(6),
-        padding: getSpacing(1),
+        width: getSpacing(switchSizeM),
+        height: getSpacing(switchSizeS),
+        padding: 0,
         overflow: 'visible',
 
         '& + .MuiFormControlLabel-label': {
           ...themeTypography.body2,
-          marginLeft: getSpacing(0.25),
+          marginLeft: getSpacing(1),
           color: commonPalette.text.primary
         }
       },
 
       switchBase: {
-        padding: getSpacing(1.5),
+        width: getSpacing(switchSizeL),
+        height: getSpacing(switchSizeL),
+        padding: getSpacing(0.5),
         borderRadius: '50%',
-        transform: 'translate(1px, 1px)',
         color: commonPalette.text.secondary,
+        transform: 'translate(-8px, -8px)',
 
+        '&:hover': {
+          backgroundColor: commonPalette.action.hover
+        },
         '&.Mui-checked': {
+          transform: 'translate(0, -8px)',
+          color: commonPalette.common.white,
+
           '& input': {
             left: getSpacing(-1.5)
           },
-
-          transform: 'translate(13px, 1px)',
-          color: commonPalette.common.white,
-
           '& + .MuiSwitch-track': {
-            opacity: 1
+            opacity: 1,
+            border: 0
           }
         }
       },
 
       thumb: {
-        width: getSpacing(1.25),
-        height: getSpacing(1.25),
-        boxShadow: 'none'
+        width: getSpacing(1),
+        height: getSpacing(1),
+        boxShadow: themeShadows[0],
+
+        '.Mui-checked &': {
+          boxShadow: themeShadows[1]
+        },
+        '.Mui-disabled &': {
+          backgroundColor: commonPalette.text.disabled
+        },
+        '.Mui-disabled.Mui-checked &': {
+          backgroundColor: commonPalette.common.white
+        }
       },
 
       input: {
-        width: getSpacing(6),
+        width: getSpacing(switchSizeM),
+        height: getSpacing(switchSizeS),
         left: 0
       },
 
       track: {
         height: 'auto',
-        border: `2px solid ${commonPalette.text.secondary}`,
+        border: `1px solid ${commonPalette.text.secondary}`,
         borderRadius: getSpacing(2),
         opacity: 1,
-        backgroundColor: commonPalette.common.white
+        backgroundColor: commonPalette.common.white,
+        transitionDuration: '300ms',
+
+        '.MuiButtonBase-root.MuiSwitch-switchBase.Mui-disabled + &': {
+          opacity: 1,
+          borderColor: commonPalette.text.disabled
+        },
+        '.MuiButtonBase-root.Mui-checked.Mui-disabled + &': {
+          backgroundColor: commonPalette.text.disabled
+        }
       },
 
       colorPrimary: {
-        '&.Mui-checked': {
-          color: commonPalette.common.white,
-
-          '& + .MuiSwitch-track': {
-            backgroundColor: commonPalette.primary.main,
-            borderColor: 'transparent'
-          },
-
-          '&.Mui-disabled': {
-            color: commonPalette.grey[100],
-
-            '& + .MuiSwitch-track': {
-              backgroundColor: commonPalette.text.disabled
-            }
-          }
-        },
-
-        '&.Mui-disabled': {
-          color: commonPalette.text.disabled,
-
-          '& + .MuiSwitch-track': {
-            opacity: 1,
-            backgroundColor: commonPalette.common.white,
-            borderColor: commonPalette.text.disabled
-          }
+        '&.Mui-checked:hover': {
+          backgroundColor: commonPalette.primary.background
         }
       },
-
       colorSecondary: {
-        '&.Mui-checked': {
-          color: commonPalette.common.white,
-
-          '& + .MuiSwitch-track': {
-            backgroundColor: commonPalette.secondary.main,
-            borderColor: 'transparent'
-          },
-
-          '&.Mui-disabled': {
-            color: commonPalette.grey[100],
-
-            '& + .MuiSwitch-track': {
-              backgroundColor: commonPalette.text.disabled
-            }
-          }
-        },
-
-        '&.Mui-disabled': {
-          color: commonPalette.text.disabled,
-
-          '& + .MuiSwitch-track': {
-            opacity: 1,
-            backgroundColor: commonPalette.common.white,
-            borderColor: commonPalette.text.disabled
-          }
-        }
-      },
-
-      sizeSmall: {
-        height: getSpacing(4.5),
-        width: getSpacing(6),
-        padding: getSpacing(1),
-
-        '& .MuiSwitch-switchBase': {
-          padding: getSpacing(1.5),
-          transform: 'translate(0, 1px)',
-
-          '&.Mui-checked': {
-            transform: 'translate(15px, 1px)'
-          }
-        },
-        '& .MuiSwitch-thumb': {
-          width: getSpacing(1.25),
-          height: getSpacing(1.25)
+        '&.Mui-checked:hover': {
+          backgroundColor: commonPalette.secondary.background
         }
       }
     }